ensured selected item is visible
This is necessary for keyboard navigation, as not all items might fit
into the viewport/container.

NB: Developers are still responsible for writing the CSS necessary to
activate a scrollable container for large result sets.

The utility function `scrollIfNecessary` will determine whether or not
an element is currently visible within its parent container, and if not,
will use the `scrollIntoView` function to scroll the element into view.
This is necessary because `Element#scrollIntoViewIfNeeded` is a
non-standard API.

Because we still want to support IE 11, we are using the boolean
parameter for `Element#scrollIntoView` because IE 11 does not support
the `#scrollIntoView({ block: "nearest" })` parameter.

Tested in:
Chrome, Firefox, Edge, Safari, IE 11 (with the help of many polyfills)