made tag name for root element configurable for simplete-suggestions
There may be instances where we want to reuse the simplete-suggestions
custom element but provide a different implementation for the simplete
form with some custom behavior. This modifies the simplete suggestions
custom element to make the selector for the root element configurable to
decouple it from the `simplete-form` tag name.