improved accessibility by adding necessary aria roles and attributes
The necessary roles are as follows:

Combobox
--------

We need `role=combobox` on the search field to let assistive
technologies know that a dropdown will open while typing.
According to the ARIA Authoring Practices Guild for Combobox, this seems
like an appropriate role for the search field:
https://www.w3.org/WAI/ARIA/apg/patterns/combobox/
https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/combobox_role

An alternative could be to set `aria-haspopup=listbox` on the search
field without the `combobox` role (a `combobox` always has a default
implicit `aria-haspopup=listbox` property).

The necessary ARIA attributes for the `combobox` are as follows:

* `aria-autocomplete` is set to `list` to make assistive technologies
  aware that there is autocomplete available in the form of a listbox
  https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-autocomplete
* `aria-expanded` must be set to `false` when the popup is closed and
  set to `true` whenever it opens -- it should never be omitted because
  the presence of `aria-expanced=false` indicates to assistive
  technologies that there is something which can be expanded.
  https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-expanded
* `aria-activedescendant` is set whenever an option is selected in the
  listbox to indicate to assistive technologies which option is
  currently selected in the dropdown.

Listbox
-------

The `search-suggestions` element receives the `listbox` role to indicate
to assistive technologies that it is a popup containing a list of
options which can be selected via the arrows keys.

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/listbox_role

Option
------

Each item that is selectable within the dropdown must receive an
`option` role. Our options correspond with the elements selected by the
`itemSelector` defined for SimpleteSuggestions.

We also need to add a generated HTML id to each option (if not already
defined) so that it can be referenced via the `aria-activedescendant`
property of the combobox.

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/option_role

For each option, we additionally need to define the following
attributes:

* `aria-selected` specifies whether or not the option is selected. Here
  we want to explicitly set `aria-selected=false` if it is not selected
  because that indicates to assistive technologies that the element is
  selectable.
  https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-selected