improved accessibility by removing focussable element from listbox
In a listbox popup, the items in the popup should only be accessible via
the arrow keys, and not via the TAB-Key. Anything else is not ideal for
accessibility or UX. A user may want to ignore the search suggestions,
and in this case, the next item in the TAB order should be the submit
button, not a list of options to select.

Additionally, by not removing the items from the TAB order, it is
possible to achieve constellations in the component which are
contradictory. For instance, if you type something into the search field
and three results come up, you can navigate with the arrow key to one of
the items. This is then highlighted and marked as selected both
visually and for assistive technologies, but then it is possible to use
the TAB key to navigate to a different link within the list. Upon
execution, the link which is selected with TAB is executed, even though
naively you would expect the item which the dropdown has marked as
selected to be executed.