improved accessibility of how the text field is filled during selection
This updates the component so that the search field will always be
populated whenever the user selects an option from within the listbox
popup. When no input field with a value is provided, the text content of
the item is used.

This is absolutely vital for accessibility because when using the
component with an assistive technology like a screenreader, the value of
the _search field_ will always be read out by the screenreader whenever
the selection in the popup changes -- that is how screenreaders expect
comboboxes to behave. Since this wasn't the case for the simplete
component, anyone using a screenreader would type in two characters into
the search field and then try to navigate the list of search suggestions
but would only hear those same two characters being read out every time
that something was suggested.

However, one thing that was noticed during the tests of this component
was that it should be possible for the user to decide to ignore the
options in the list and continue typing to complete their original
intended search string. This has been implemented in this commit by
checking if the user key typed is alphanumeric, and if so, restoring the
original query and aborting the simplete popup. It's possible that we
should extend the number of "normal user input" keys in the future to
recognize more than alphanumeric characters.

Note that the keyboard behavior has been tested in the following
browsers:

* IE11
* Firefox 45.3.0
* Modern Firefox
* Chrome
* Safari
* Edge