added simplete-status for announcing current status to assistive tech
This adds a simplete-status custom element which we can use as an ARIA
live region to announce to the user what the current status of the
component is (most often this will be something like: "5 search
suggestions available"). The simplete-status component will
automatically be created within the simplete-form.

The actual autocomplete status is implemented by passing the current
status via the `data-autocomplete-status` in the result HTML. This means
that i18n must be implemented by the server.

One "bug" that was found in the first version of this implementation is
that the simplete-status element did not disappear by default when the
focus left the input field and was still available as a text node when
navigating through the DOM with assistive technologies. This also adds
some CSS `label:not(:focus-within) + simplete-status, input:not(:focus)
+ simplete-status { display: none; }` which will hide the status field
correctly if placed directly behind the search field in the HTML DOM.
The css works for both an `input` field wrapped in a `label` (at least
for browsers which support `:focus-within`), and for an `input` field
which is referenced by its label via its `for` attribute.