simplified custom-element constructors
The `self` parameter was added to the constructor of `SimpleteForm` and
`SimpleteSuggestions` to account for the `document-register-element`
polyfill. That is no longer the recommended polyfill for custom elements
though, having been deprecated in favor of `@ungap/custom-elements`.

The code has been tested with the `@ungap/custom-elements` polyfill and
works without the `self` parameter. Note that the component in IE 11 had
problems with the response from `demo/servers/languages.js` when the
HTML response consisted of two HTML nodes instead of one, so this also
wraps the response in a `<div>` in order to ensure that only one HTML
node is returned to the client.

This change may warrant a major release.