added simplete abort when the user continues typing
This adds a case for when the user is navigating in the dropdown of
simplete results. If the user does not find a result that they are
interested in, they can continue typing and then their original input
will be restored and the simplete will be aborted. This is especially
necessary for assistive technologies because they will likely not be
able to inspect the results in the simplete list without inspecting each
one of them.

This also refactors the implementation of `this.navigating`. Instead of
using a variable that we have to explicitly set, we can check if the
`aria-activedescendant` attribute is set on the search field, because if
this is set, we know that the user is currently navigating within the
simplete dropdown. When the dropdown is toggled, the
`aria-activedescendant` attribute is always removed.

Note that this also slightly tweaks the `onConfirm` method within
SearchSuggestions which requires a review because I'm not completely
clear on how the onConfirm method should behave.