refactored navigating variable
Upon review of the code, it seems that `preview` referred to the act of
viewing a suggestion without actually selecting it (e.g. it controlled
whether or not the dropdown window would close upon selection).

This is exactly the meaning of `navigating` in the `SimpleteForm`
component, so this refactors the `SimpleteSuggestions` component to use
a property named `navigating` instead of `preview`.

This also restructures the code to use memoization of the navigating
state for performance reasons.